Restructuring things for long term cleanliness
